test to delete and link accounts following https://firebase.google.com/docs/auth/web/account-linking
In case of email link it seems to not be working because the methods
requires us to login is using the email token that we received, to delete
the user, to link the email token we received with the new account and
then login again using the current credentials.
The problem is that the credentials are emaillink credentials and they
expire immediately after they are used once so reconnect is not possible.